+++ lib/sugar_refinery/all.rb
@@ -1,2 +1,5 @@
-require 'sugar_refinery'
-SugarRefinery.require_all
+require_relative 'version'
+
+Dir["#{File.dirname(__FILE__)}/*"].each{ |f|
+  require_relative File.basename(f)
+}