+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Sugar Refinery Changelog
 
 ## 2016-01-01 | Sugar Refinery 1.0
-* Turn zucker gem into sugar_refinery
-* Remove 'Zucker#require' and 'Zucker#require_all'
-* Remove 'refine' gem dependency
+* turn zucker gem into sugar_refinery
+* remove 'Zucker#require' and 'Zucker#require_all'
+* remove 'refine' gem dependency
+* remove square_brackets_for
 
 
 ## CHANGELOG FOR ZUCKER GEM